updated scripts for axelar-core v0.7.5 and tofnd v0.6.1 (#139)

+++ join/consumeGenesis.sh
@@ -1,17 +1,30 @@
 #!/bin/sh
-set -e
+export NODE_MONIKER=${NODE_MONIKER:-"$(hostname)"}
+export AXELARD_CHAIN_ID=${AXELARD_CHAIN_ID:-"axelar"}
 
-axelard init "$1" --chain-id "$2"
-if [ -f "$TENDERMINT_KEY" ]; then
-  cp -f "$TENDERMINT_KEY" "/root/.axelar/config/priv_validator_key.json"
-fi
+ACCOUNTS=$(axelard keys list -n)
+for ACCOUNT in $ACCOUNTS; do
+    if [ "$ACCOUNT" == "validator" ]; then
+        HAS_VALIDATOR=true
+    fi
+done
 
-if [ -f "$AXELAR_MNEMONIC" ]; then
-  axelard keys add validator --recover <"$AXELAR_MNEMONIC"
-  touch "/validator.txt"
-else
-  axelard keys add validator >"/validator.txt" 2>&1
+touch "/validator.txt"
+if [ -z "$HAS_VALIDATOR" ]; then
+  if [ -f "$AXELAR_MNEMONIC" ]; then
+    axelard keys add validator --recover <"$AXELAR_MNEMONIC"
+  else
+    axelard keys add validator >"/validator.txt" 2>&1
+  fi
 fi
 
 axelard keys show validator -a --bech val > "/root/shared/validator.bech"
-cp "/root/shared/genesis.json" "/root/.axelar/config/genesis.json"
+
+if [ ! -f "/root/.axelar/config/genesis.json" ]; then
+  axelard init "$NODE_MONIKER" --chain-id "$AXELARD_CHAIN_ID"
+  if [ -f "$TENDERMINT_KEY" ]; then
+    cp -f "$TENDERMINT_KEY" "/root/.axelar/config/priv_validator_key.json"
+  fi
+
+  cp "/root/shared/genesis.json" "/root/.axelar/config/genesis.json"
+fi
